[lldb] Fix a warning
This patch fixes:

  lldb/tools/lldb-dap/DAP.cpp:975:34: error: format specifies type
  'long long' but the argument has type 'rep' (aka 'long')
  [-Werror,-Wformat]

+++ lldb/tools/lldb-dap/DAP.cpp
@@ -971,8 +971,9 @@ lldb::SBError DAP::WaitForProcessToStop(std::chrono::seconds seconds) {
     }
     std::this_thread::sleep_for(std::chrono::microseconds(250));
   }
-  error.SetErrorStringWithFormat("process failed to stop within %lld seconds",
-                                 seconds.count());
+  error.SetErrorStringWithFormat("process failed to stop within %" PRId64
+                                 " seconds",
+                                 static_cast<uint64_t>(seconds.count()));
   return error;
 }
